requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.9.2
alembic==1.13.2